% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTypes.R
\name{ColSpec2ShortType}
\alias{ColSpec2ShortType}
\title{ColSpec2ShortType}
\usage{
ColSpec2ShortType(colspec)
}
\arguments{
\item{colspec}{A \link[vroom:cols]{vroom::cols} specification}
}
\value{
a character vector with single characters
}
\description{
Converts a \link[vroom:cols]{vroom::cols} specification and turns it into standard abbreviations.
}
\examples{
ColSpec2ShortType(vroom::cols(a = "i", b = "d", c = "T"))
#> integer  double datetime
#>     "i"     "d"      "T"
}
