% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColumnConverter.R
\name{ColumnConvert}
\alias{ColumnConvert}
\title{Convert columns of a data frame from character to their desired type}
\usage{
ColumnConvert(Columns, Converter, Format)
}
\arguments{
\item{Columns}{A data frame with columns to cast types}

\item{Converter}{A list (see details)}

\item{Format}{A list of formats (see details)}
}
\description{
Convert columns of a data frame from character to their desired type
}
\details{
Valid converters: character, logical, numeric, double, integer, factor, datetime, find,
}
