% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Converter.R
\name{DataFrameConvert}
\alias{DataFrameConvert}
\title{Change the type of data frame columns.}
\usage{
DataFrameConvert(Df, ColSpec, FileSpec, Preview = FALSE)
}
\arguments{
\item{Df}{A data frame}

\item{ColSpec}{A list specifying the columns to import.}

\item{FileSpec}{A list describing the expected CSV format.
(see details).}

\item{Preview}{FALSE (default) returns a working data frame. If \code{Preview} is \code{TRUE} or
numeric all columns are character. The number of rows can be changed with \code{Preview} being
numeric.}
}
\value{
A data frame
}
\description{
Change the type of data frame columns.
}
