% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{.HandleUTF8}
\alias{.HandleUTF8}
\title{Convert characters to utf8 representation}
\usage{
.HandleUTF8(x)
}
\arguments{
\item{x}{A string}
}
\value{
All characters are replaced by unicode representations
}
\description{
Convert characters to utf8 representation
}
