% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreviewRenderHelper.R
\name{.renderTableRow}
\alias{.renderTableRow}
\title{.renderTableRow}
\usage{
.renderTableRow(Elements)
}
\arguments{
\item{Elements}{A list of cell contents. The created row will have
as many cells as \code{Elements} has entries.}
}
\value{
A \code{list()} with a \code{shiny.tag} class.
}
\description{
Render elements of a preview table
}
\seealso{
This is a helper function for the \link{renderRowFunctions}.
}
