% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindLocale.R
\name{FindLocale}
\alias{FindLocale}
\title{Find a locale specification for a given language code}
\usage{
FindLocale(LangCode)
}
\arguments{
\item{LangCode}{A language code like "de-CH"  or "en-GB"
(i.e. in the form of a IETF language tag).}
}
\value{
A locale object (see readr::locale())
}
\description{
Find a locale specification for a given language code
}
\references{
\url{https://en.wikipedia.org/wiki/IETF_language_tag},
\url{https://cran.r-project.org/web/packages/readr/vignettes/locales.html}
}
\seealso{
\code{\link[readr:locale]{readr::locale()}} where \code{locale} objects are specified
}
