% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LanguageCodes.R
\name{FormatLocaleCode}
\alias{FormatLocaleCode}
\title{Fixes typical inconsistencies in the use of language codes}
\usage{
FormatLocaleCode(Code)
}
\arguments{
\item{Code}{A code as string in the format "xx-XX" or just "xx".}
}
\description{
This function replaces '_' characters with '-'. Primary language codes are
forced to lower case and country codes are forced to upper case letters.
}
