% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LanguageCodes.R
\name{GetLanguageTag}
\alias{GetLanguageTag}
\title{Extracts the primary language subtags from a list of language tags}
\usage{
GetLanguageTag(Code)
}
\arguments{
\item{Code}{A code as string in the format "xx-XX" or just "xx".}
}
\value{
the second subtag(s) extracted from the given codes.
}
\description{
Extracts the primary language subtags from a list of language tags
}
\note{
The function does not verify if the country code exists
}
