% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Converter.R
\name{GuessColumnTypes}
\alias{GuessColumnTypes}
\title{Guess all column types of a data frame from character data}
\usage{
GuessColumnTypes(Data, Locale = "de-DE")
}
\arguments{
\item{Data}{A data frame with all columns being of type \code{character}.}

\item{Locale}{Information about the underlying locale}
}
\value{
A vector of strings indicating the data types
}
\description{
Guess all column types of a data frame from character data
}
\seealso{
\code{\link[readr:parse_guess]{readr::guess_parser()}}, \code{\link[readr:locale]{readr::locale()}}
}
