% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeekIntoFile.R
\name{HasFileHeader}
\alias{HasFileHeader}
\title{HasFileHeader}
\usage{
HasFileHeader(HeaderFileSpec, File, ...)
}
\arguments{
\item{HeaderFileSpec}{A specification read from a CSV file with \code{col_names = TRUE}, i.e.
with the assumption that the file has a header.}

\item{File}{The according file that made the data set of \code{Data}.}

\item{...}{Additional arguments handed over to \code{\link[vroom:vroom]{vroom::vroom}}.}
}
\value{
TRUE/FALSE
}
\description{
HasFileHeader
}
