% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreviewRenderHelper.R
\name{inlinewidgets}
\alias{inlinewidgets}
\alias{.renderTextInput}
\alias{.renderCheckBox}
\alias{.renderSelect}
\title{Inline Widgets}
\usage{
.renderTextInput(.colname, .label, .val, .enable)

.renderCheckBox(.colname, .label, .val, .enable)

.renderSelect(.colname, .label, .val, .chc, .enable)
}
\value{
A \code{list()} with a \code{shiny.tag} class.
}
\description{
Each function creates a tag list of interactive widgets fitting into a table.
}
\details{
To fit it into the table the code created by Shiny has been adapted (at the cost of
a dependency to the current Shiny version this package was created for).

\code{.renderTextInput} renders a single test input field to be used inline.
It generates a modified code of this:

\if{html}{\out{<div class="sourceCode">}}\preformatted{<div class="form-group shiny-input-container">
  <label class="control-label" id="ColName 2-label" for="ColName 2">...</label>
  <input id="ColName 2" type="text" class="form-control" value="A"/>
</div>
}\if{html}{\out{</div>}}

In this case we 1. skip the label, 2. remove the class "form-group" because it
just adds a bottom padding, 3.

\code{.renderCheckBox} renders a single check box to be used inline.

\code{.renderSelect} renders a select input to be used inline.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# The original Shiny code in 1.7.1 will be modified for inline use
<div class="form-group shiny-input-container">
  <label class="control-label" id="id-label" for="id">label</label>
  <div>
    <select id="id">
       <option value="A" selected>A</option>
     </select>
     <script type="application/json" data-for="id" data-nonempty="">\{"plugins":["selectize-plugin-a11y"]\}</script>
   </div>
 </div>
}\if{html}{\out{</div>}}
}
\seealso{
These are helper functions for the \link{renderRowFunctions}.
}
