% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TruthyHelpers.R
\name{isTruthyInside}
\alias{isTruthyInside}
\title{isTruthyInside}
\usage{
isTruthyInside(x)
}
\arguments{
\item{x}{A vector or list}
}
\value{
A logical vector for each position indicating \code{TRUE}/\code{FALSE}.
Returns \code{NULL} if the object \code{x} itself is not truthy or missing.
}
\description{
\code{isTruthyInside} tests the "truthiness" of positions
of a vector or list (while \code{isTruthy} tests the vector as a whole).
}
\details{
\code{isTruthyInside}
}
\examples{
isTruthyInside(1:5)
#> [1] TRUE TRUE TRUE TRUE TRUE
isTruthyInside(c(1, NA, 2))
#> [1]  TRUE FALSE  TRUE
isTruthyInside(list(1, NA, NULL, integer(0), 2))
#> [1]  TRUE FALSE FALSE FALSE  TRUE
}
