% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportModule.R
\name{ModuleImportUI}
\alias{ModuleImportUI}
\title{\code{ModuleImportUI} is the UI function of the CSV import module}
\usage{
ModuleImportUI(id)
}
\arguments{
\item{id}{Module namespace to be set by by caller}
}
\description{
\code{ModuleImportUI} is the UI function of the CSV import module
}
