% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeekIntoFile.R
\name{PeekIntoFile}
\alias{PeekIntoFile}
\title{PeekIntoFile
Investigates the format of a delimiter separated file.}
\usage{
PeekIntoFile(File, GetSpec = FALSE, ...)
}
\arguments{
\item{File}{The name of the file to investigate}

\item{GetSpec}{Shall the function return a \code{col_spec} object or
the data?}

\item{...}{Additional arguments handed over to \code{\link[vroom:vroom]{vroom::vroom}}.}
}
\value{
Depending on \code{GetSpec} the column specification or the
data frame (as returned from \link{vroom}).
}
\description{
PeekIntoFile
Investigates the format of a delimiter separated file.
}
