% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreviewRenderHelper.R
\name{renderRowFunctions}
\alias{renderRowFunctions}
\alias{renderRowTextInput}
\alias{renderRowCheckBox}
\title{renderRow...-Functions}
\usage{
renderRowTextInput(ColNames, Label = NULL, Values = NULL, Enabled = TRUE)

renderRowCheckBox(ColNames, Label = NULL, Values = NULL, Enabled = TRUE)
}
\arguments{
\item{ColNames}{The names of the columns are required to generate input-ids
for the widgets.}

\item{Label}{An optional label for the input widget in each column.
Not supported for all interactive row types.}

\item{Values}{Optional values to be set.}

\item{Enabled}{Is the row interactive or does it only show the current
state (not vectorised).}
}
\value{
A \code{list()} with a \code{shiny.tag} class that can be converted into an HTML
string via \code{as.character()} and saved to a file with \code{\link[=save_html]{save_html()}}.
}
\description{
These functions render an \strong{interactive row} of the preview table used
to import CSV files in the module.
}
\details{
\code{renderRowCheckBox} creates a tag list of a table row with a check box in each
cell.

\code{renderRowCheckBox} creates a tag list of a table row with a check box in each
cell.
}
\seealso{
About Shiny HTML tag lists refer to \link[htmltools:builder]{htmltools::builder}.
}
