% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run shiny.CSVImport Example Applications}
\usage{
runExample(
  example = NA,
  port = getOption("shiny.port"),
  launch.browser = getOption("shiny.launch.browser", interactive()),
  host = getOption("shiny.host", "127.0.0.1"),
  display.mode = c("auto", "normal", "showcase")
)
}
\arguments{
\item{example}{The name of the example to run, or \code{NA} (the default) to
list the available examples.}

\item{port}{The TCP port that the application should listen on. If the
\code{port} is not specified, and the \code{shiny.port} option is set (with
\code{options(shiny.port = XX)}), then that port will be used. Otherwise,
use a random port.}

\item{launch.browser}{If true, the system's default web browser will be
launched automatically after the app is started. Defaults to true in
interactive sessions only.}

\item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not.}

\item{display.mode}{The mode in which to display the example. Defaults to
\code{showcase}, but may be set to \code{normal} to see the example without
code or commentary.}
}
\description{
Launch shiny.CSVImport example applications, and optionally, your system's web browser.
}
\examples{
# List all available examples
# (be sure to use the name space prefix to avoid conflicts)
shiny.CSVImport::runExample()

## Only run this example in interactive R sessions
if (interactive()) {
  # Run one of the examples
  shiny.CSVImport::runExample("01_hello")
}

# Find the directory containing the code for all examples
system.file("examples", package="shiny.CSVImport")
}
