% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.CSVImport-package.R
\docType{package}
\name{shiny.CSVImport-package}
\alias{shiny.CSVImport}
\alias{shiny.CSVImport-package}
\title{shiny.CSVImport: Easily import CSV files into a Shiny App with a pre-built import module}
\description{
A package providing a module with all functionality needed to import CSV files to a Shiny app. The module allows the user to make the required settings (like decimal and column separator character). The module returns a data frame.
}
\details{
A package supporting interactive import of CSV (DSV) files in a shiny app.
The package provides a module with an UI and server implementation. Developers simply
include the package. Users can change the locale settings and see the effect of various
settings in a preview. That way, they can easily import CSV files from various sources.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SigurdJanson/shinyCSVImpoMod}
  \item Report bugs at \url{https://github.com/SigurdJanson/shinyCSVImpoMod/issues}
}

}
\author{
\strong{Maintainer}: Jan Seifert \email{github@sigurdjanson.anonaddy.com}

}
\keyword{internal}
