% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VerifyMode.R
\name{VerifyMode}
\alias{VerifyMode}
\title{VerifyMode
Tests and verifies the argument \code{Mode} of the module server function
\code{ModuleImportServer}.}
\usage{
VerifyMode(Mode)
}
\arguments{
\item{Mode}{The interaction mode for the CSV import module.}
}
\value{
If successfully verified the function returns the correct mode
(see also \code{\link[=match.arg]{match.arg()}}).
}
\description{
VerifyMode
Tests and verifies the argument \code{Mode} of the module server function
\code{ModuleImportServer}.
}
\details{
matches the argument against it's default values.
}
