% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Converter.R
\name{ColumnConvert}
\alias{ColumnConvert}
\title{ColumnConvert}
\usage{
ColumnConvert(Data, Converter, Format, Locale)
}
\arguments{
\item{Data}{A data frame with columns to cast types}

\item{Converter}{A list (see details)}

\item{Format}{A list of formats (see details)}

\item{Locale}{A locale that can be used in case format is missing}
}
\value{
a data frame with changed column data types
}
\description{
Convert columns of a data frame to their desired types. Columns in
source data frame are all of type \code{character}.
}
\details{
Valid converters: date, time, datetime, character, factor, logical, number,
double, integer. Long and short names are allowed (see \code{\link[vroom:cols]{vroom::cols()}}.

\code{NULL} drops a column. Everything else will be  guessed by \code{\link[readr:parse_guess]{readr::parse_guess()}}.

The types date, time, and datetime \strong{support} a \code{Format} specification.
If none is available the \code{Locale} is used.
}
