% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TruthyHelpers.R
\name{isTruthyInside}
\alias{isTruthyInside}
\title{isTruthyInside tests the truthiness of positions of a vector (while \code{isTruthy}
tests the vector as a whole)}
\usage{
isTruthyInside(x)
}
\arguments{
\item{x}{A vector or list}
}
\value{
A logical vector for each position indicating \code{TRUE}/\code{FALSE}.
Returns \code{NULL} if operation was not successful.
}
\description{
isTruthyInside tests the truthiness of positions of a vector (while \code{isTruthy}
tests the vector as a whole)
}
\examples{
isTruthyInside(1:5)
#> [1] TRUE TRUE TRUE TRUE TRUE
isTruthyInside(c(1, NA, 2))
#> [1]  TRUE FALSE  TRUE
isTruthyInside(list(1, NA, NULL, integer(0), 2))
#> [1]  TRUE FALSE FALSE FALSE  TRUE
}
