% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportModule.R
\name{ModuleImportServer}
\alias{ModuleImportServer}
\title{ModuleImportServer - the server function of the CSV import module}
\usage{
ModuleImportServer(id, stringsAsFactors = TRUE)
}
\arguments{
\item{id}{Module namespace}

\item{stringsAsFactors}{Shall strings be converted to factors (unless specified otherwise) (\code{TRUE}/\code{FALSE}).}
}
\description{
ModuleImportServer - the server function of the CSV import module
}
