% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportModule.R
\name{ModuleImportUI}
\alias{ModuleImportUI}
\title{The UI function of the CSV import module}
\usage{
ModuleImportUI(Id)
}
\arguments{
\item{Id}{Module namespace to be set by caller}
}
\description{
The UI function of the CSV import module
}
