% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TruthyHelpers.R
\name{PickTruthy}
\alias{PickTruthy}
\title{Choose a value if it truthy. If not, choose it's replacement}
\usage{
PickTruthy(X, Replacement)
}
\arguments{
\item{X}{The designated value}

\item{Replacement}{The replacement in case \code{X} is falsy}
}
\value{
\code{X} or \code{Replacement}
}
\description{
Choose a value if it truthy. If not, choose it's replacement
}
\examples{
PickTruthy(1, 2) # 1
PickTruthy("1", 2) # "1"
PickTruthy("", 2) # 2
PickTruthy(NULL, 3) # 3
PickTruthy(NULL, NA) # NA
}
