% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VerifyColSpecFormat.R
\name{VerifyColSpecFormat}
\alias{VerifyColSpecFormat}
\title{VerifyColSpecFormat}
\usage{
VerifyColSpecFormat(ColSpec)
}
\arguments{
\item{ColSpec}{A column specification. Either an object of class \code{col_spec}}
}
\value{
A column specification in the format of a \code{col_spec} class.
}
\description{
VerifyColSpecFormat
}
\seealso{
\code{\link[readr:spec]{readr::cols_condense()}}
}
